1в 12 Т3

+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/UserStorage.java
@@ -1,16 +1,16 @@
 package ru.yandex.practicum.filmorate.storage.user;
 
-import ru.yandex.practicum.filmorate.model.User;
+import ru.yandex.practicum.filmorate.model.users.User;
 
 import java.util.Collection;
 
 public interface UserStorage {
-
     Collection<User> getAllUsers();
 
     User addNewUser(User user);
 
     User updateUser(User updatedUser);
 
-    User findUserById(Long userId);
+    User findUserById(Integer userId);
+
 }
